방명록, 즐겨찾기, 로그인, 로그아웃 페이지 추가 및 수정

+++ dynamic_web01/src/main/java/controller/BookmarkDeleteController.java
@@ -6,21 +6,32 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import model.dto.BookmarkDTO;
+import model.dto.UserDTO;
 import model.service.BookmarkService;
 
 public class BookmarkDeleteController extends HttpServlet {
 
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		String id = req.getParameter("id");
+		HttpSession session = req.getSession();
 
+		if(session.getAttribute("login") == null) {
+			resp.sendRedirect(req.getContextPath() + "/login");
+			return;
+		}
+
+		UserDTO userData = (UserDTO)session.getAttribute("user");
+
+		
 		BookmarkDTO dto = new BookmarkDTO();
-		dto.setId(Integer.parseInt(id));
-
+		dto.setUserId(userData.getUserId());
+		dto.setUserId(userData.getUserId());
+		
 		BookmarkService service = new BookmarkService();
 		boolean result = service.remove(dto);
-
+		
 		if(result) {
 			resp.sendRedirect("../bookmark");
 		} else {
